BluetoothFoundAdapter can now respond to recycler items clicks
removed button form bluetooth_found.
BluetoothFoundAdapter now has an interface that is called when a recycler item is pressed.
BluetoothFoundAdapter can set a listener for the new interface
interface method is used to pass the mac adressa and device name to the activity.
each view holder calls the listeners method if it was selected
toast is made that displays the selected items name and address.