Pairing and ServerSetup activity now start a new Chatroom Activity when they have been connected with a valid bluetooth socket. Only the basic operations needed in order to start the chat room activity have been implemented.
Both bind to BluetoothRead and BluetoothWrite services in on create and unbind in onDestroy.

Once a client connects to a server, it gives the socket to the services and starts the chat room.
Once the Server gets a request from the client, it gives the socket to the services and starts the chat room.

ChatroomActivity.
     - User types text and sends it to the other using socket
     - Very incomplete layout and implementation since next focus will be  improving
        the users experience before chatroom activity is started.