Removed both services dependencies on the utility class.
This was done so that the service can be used by other applications without having to
modify the service code.

BluetoothReadService now takes a handler as a parameter rather than a ReadServiceHandler.
This was also done so that the service can work with any handler.

Utility now has the method sendDisplayText that used to be in BluetoothWriteService